requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
playwright
fake-useragent
pdfplumber
python-multipart
spacy
huggingface_hub
sentence-transformers
numpy
sqlite-vec
langchain
langchain-groq
langchain-core
pywebview
